+++ editoast/src/models/sql/generate_speed_section_layer.sql
@@ -2,10 +2,10 @@ WITH track_ranges AS (
     SELECT obj_id as speed_id,
         (
             jsonb_array_elements(data->'track_ranges')->'begin'
-        )::numeric as slice_begin,
+        )::float as slice_begin,
         (
             jsonb_array_elements(data->'track_ranges')->'end'
-        )::numeric as slice_end,
+        )::float as slice_end,
         jsonb_array_elements(data->'track_ranges')->'track'->>'id' as track_id
     FROM osrd_infra_speedsectionmodel
     WHERE infra_id = $1
@@ -16,11 +16,11 @@ sliced_tracks AS (
             ST_LineSubstring(
                 ST_GeomFromGeoJSON(tracks.data->'geo'),
                 GREATEST(
-                    track_ranges.slice_begin / (tracks.data->'length')::numeric,
+                    track_ranges.slice_begin / (tracks.data->'length')::float,
                     0.
                 ),
                 LEAST(
-                    track_ranges.slice_end / (tracks.data->'length')::numeric,
+                    track_ranges.slice_end / (tracks.data->'length')::float,
                     1.
                 )
             ),
@@ -30,11 +30,11 @@ sliced_tracks AS (
             ST_LineSubstring(
                 ST_GeomFromGeoJSON(tracks.data->'sch'),
                 GREATEST(
-                    track_ranges.slice_begin / (tracks.data->'length')::numeric,
+                    track_ranges.slice_begin / (tracks.data->'length')::float,
                     0.
                 ),
                 LEAST(
-                    track_ranges.slice_end / (tracks.data->'length')::numeric,
+                    track_ranges.slice_end / (tracks.data->'length')::float,
                     1.
                 )
             ),
